extract battle body
