+++ modules/ui/src/components/GamePage/GamePage.tsx
@@ -2,28 +2,33 @@
  * Used in the Next app as a "homepage" for the individual games.
  */
 
-import React, { ReactNode } from "react";
+import React from "react";
 
-import { Board } from "../Board";
 import { Page } from "../Page";
-import { BattleLanding } from "../BattleLanding";
-import { GameLanding } from "../GameLanding";
-import { BattleHistory } from "../BattleHistory";
 import { useUI } from "./GamePage.useUI";
-import { BattleMove } from "../BattleMove";
-import { getLatestSessionIdForGame } from "../../../../local/expose";
 import { useAppState, useGamePayload } from "../../contexts";
 import { useBattleActionsAndState } from "./GamePage.useBattleActionsAndState";
 import { useBattleEffects } from "./GamePage.useBattleEffects";
+import { Board } from "../Board";
+import { GamePageBody } from "./GamePage.Body";
 
 export const GamePage = () => {
+  // static assets for the current game
   const gamePayload = useGamePayload();
+  const { api, graphics, meta, demo } = gamePayload;
+
+  // current state of the app (URL parameters)
   const { battleMode: givenMode = "gamelobby", sessionId } = useAppState();
-  const { api, graphics, meta, demo, rules } = gamePayload;
+
+  // current battle state
   const [state, battleActions] = useBattleActionsAndState(api);
-  const { battle, frame, session } = state;
-  const ui = useUI(api, battle, frame, demo, givenMode);
+  const { battle, frame } = state;
   const mode = battle ? givenMode : "gamelobby";
+
+  // current rendered board
+  const ui = useUI(api, battle, frame, demo, givenMode);
+
+  // battle-related side effects
   useBattleEffects({
     api,
     actions: battleActions,
@@ -33,45 +38,6 @@ export const GamePage = () => {
     sessionId,
     ui,
   });
-
-  // TODO - maybe not read this on every render? move to state somewhere?
-  const previousSessionId = getLatestSessionIdForGame(api.gameId);
-
-  let body: ReactNode;
-  if (mode === "history") {
-    // We are currently watching the history of a battle
-    body = (
-      <BattleHistory
-        content={ui.instruction}
-        frame={Math.max(0, frame)}
-        toFrame={battleActions.toFrame}
-        battle={battle!}
-      />
-    );
-  } else if (mode === "battlelobby") {
-    // we are on the landing page for a created/loaded session
-    body = (
-      <BattleLanding
-        battle={battle!}
-        session={session!}
-        meta={meta}
-        manyVariants={api.variants.length > 1}
-      />
-    );
-  } else if (mode === "playing") {
-    // We are actively playing an ongoing battle
-    body = <BattleMove actions={battleActions} ui={ui} rules={rules} />;
-  } else {
-    // No battle active, we're just at the game landing page
-    body = (
-      <GameLanding
-        meta={meta}
-        graphics={graphics}
-        previousSessionId={previousSessionId}
-        variants={api.variants}
-      />
-    );
-  }
 
   return (
     <Page
@@ -88,7 +54,15 @@ export const GamePage = () => {
           name={mode !== "gamelobby" ? battle!.variant.board : "basic"}
         />
       }
-      body={body}
+      body={
+        <GamePageBody
+          ui={ui}
+          mode={mode}
+          battleActions={battleActions}
+          battleState={state}
+          gamePayload={gamePayload}
+        />
+      }
     />
   );
 };
